+++ src/@types/createLayers.types.ts
@@ -16,15 +16,5 @@ interface IDependencies {
   service: string[];
   factory: string[];
 }
-interface ITemplates {
-  [key: string]: (componentName: string, repositoryName?: string, serviceName?: string) => Template;
-  repositoryTemplate: (componentName: string) => Template;
-  serviceTemplate: (componentName: string, repositoryName: string) => Template;
-  factoryTemplate: (componentName: string, serviceName: string) => Template;
-}
-interface Template {
-  fileName: string
-  template: string
-}
 
-export { ICreateLayersProps, ICreateFilesProps, IDependencies, ITemplates };
+export { ICreateLayersProps, ICreateFilesProps, IDependencies };
